changed project structure to fix circular dependency #1447 (#1453)
* changed project structure to fix circular dependency
- moved functions dependent on 'sagaHelpers' imports to a separate file
- added second export to 'effects' for new file/functions

* renamed helpers file

* updated path to match name change
